medisafe-v2/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
